test(canon): expect the authored specifier in the namespace CLI check
The CLI regression test still pinned the mirror '.vue.ts' spelling in its TS2614 expectation, and the module-specifier extraction pushed diagnostics.rs one line past its length budget; the TS2307 message-rewrite test now lives in the module that owns the rewrite.
